+++ src/FindDirectories.py
@@ -1,0 +1,9 @@
+import os
+
+#Takes a path, and gets every file in that folder, recursively.
+def getAllFiles(path):
+    listofFiles = []
+    for root, dirs, files in os.walk(path):
+        for file in files:
+            listofFiles.append((os.path.join(root,file))[len(path):])
+    return listofFiles
